# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ContentCloud Private Transport
  version: 1.0.0
  description: >-
    Private same-origin Web BFF, public review BFF, and CLI transport. This is not
    a public SDK. Agents, renderers, scripts, and CI use the contentcloud CLI.
    The cloud control plane never accepts LLM prompts, model credentials, or
    arbitrary execution requests.
servers:
  - url: /api
...
paths:
...
  /bff/projects/{project_id}/connector-bindings:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    get:
      tags: [Web BFF]
      operationId: listConnectorBindings
      security: [{cookieAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
    post:
      tags: [Web BFF]
      operationId: createConnectorBinding
      summary: Bind a connector with an opaque SecretRef, never plaintext credentials
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/CreateConnectorBindingInput'}}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
...
  /bff/projects/{project_id}/channel-bindings:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    get:
      tags: [Web BFF]
      operationId: listChannelBindings
      security: [{cookieAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
    post:
      tags: [Web BFF]
      operationId: createChannelBinding
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/CreateChannelBindingInput'}}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /bff/channel-publications:
    get:
      tags: [Web BFF]
      operationId: listChannelPublications
      security: [{cookieAuth: []}]
      parameters:
        - {name: task_id, in: query, required: false, schema: {type: string, format: uuid}}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
    post:
      tags: [Web BFF]
      operationId: prepareChannelPublication
      summary: Freeze a DeliveryPackage into an idempotent publication intent
      security: [{cookieAuth: []}]
      parameters:
        - {name: Idempotency-Key, in: header, required: false, schema: {type: string, maxLength: 128}}
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/PrepareChannelPublicationInput'}}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /bff/channel-publications/reconcile:
    post:
      tags: [Web BFF]
      operationId: reconcileChannelPublications
      summary: Inspect pending publications and import authoritative external states
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/ReconcileChannelPublicationsInput'}}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
  /bff/channel-publications/{id}/submit:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post:
      tags: [Web BFF]
      operationId: submitChannelPublication
      security: [{cookieAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /bff/channel-publications/{id}/inspect:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post:
      tags: [Web BFF]
      operationId: inspectChannelPublication
      security: [{cookieAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /bff/channel-publications/{id}/receipt:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post:
      tags: [Web BFF]
      operationId: recordManualChannelReceipt
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/RecordManualChannelReceiptInput'}}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /bff/channel-publications/{id}/withdraw:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post:
      tags: [Web BFF]
      operationId: withdrawChannelPublication
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason: {type: string, minLength: 1}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    cookieAuth: {type: apiKey, in: cookie, name: cc_session}
    bearerAuth: {type: http, scheme: bearer}
  parameters:
    ProjectID: {name: project_id, in: path, required: true, schema: {type: string, format: uuid}}
    ID: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
  schemas:
...
    CreateConnectorBindingInput:
      type: object
      additionalProperties: false
      required: [connector_id, authorization_ref]
      properties:
        connector_id: {type: string, minLength: 1}
        authorization_ref: {type: string, pattern: '^(secret|vault|env)://.+'}
        region: {type: string, default: global}
...
    CreateChannelBindingInput:
      type: object
      additionalProperties: false
      required: [channel, adapter_id, account_ref, authorization_secret_ref]
      properties:
        channel: {type: string, minLength: 1}
        adapter_id: {type: string, minLength: 1, default: manual}
        account_ref: {type: string, minLength: 1}
        authorization_secret_ref: {type: string, pattern: '^(secret|vault|env)://.+'}
        region: {type: string, default: global}
...
    PrepareChannelPublicationInput:
      type: object
      additionalProperties: false
      required: [task_delivery_id, channel_binding_id]
      properties:
        task_delivery_id: {type: string, format: uuid}
        channel_binding_id: {type: string, format: uuid}
        idempotency_key: {type: string, maxLength: 128}
        content_profile_id: {type: string, enum: [douyin-commerce-video, wechat-official-article, serialized-novel]}
        douyin_commerce: {$ref: '#/components/schemas/DouyinCommercePublicationRefs'}
        scheduled_at: {type: string, format: date-time}
        metadata: {type: object, additionalProperties: true}
    DouyinCommerceOfferFacts:
      type: object
      additionalProperties: false
      required: [sku_id, product_version_id, display_price, currency, benefits, conditions]
      properties:
        sku_id: {type: string, minLength: 1}
        product_version_id: {type: string, minLength: 1}
        display_price: {type: string, minLength: 1}
        currency: {type: string, pattern: '^[A-Z]{3}$'}
        benefits: {type: array, uniqueItems: true, items: {type: string, minLength: 1}}
        conditions: {type: array, uniqueItems: true, items: {type: string, minLength: 1}}
    DouyinCommerceValidationReceipt:
      type: object
      additionalProperties: false
      required: [schema_version, content_profile_id, project_id, audience_strategy_approved_snapshot_id, audience_strategy_version_id, offer_approved_snapshot_id, offer_snapshot_id, content_approved_snapshot_id, content_item_id, content_item_digest, storyboard_approved_snapshot_id, storyboard_package_id, storyboard_locked_digest, rendered_creative_artifact_id, rendered_creative_digest, voiceover_text_digest, on_screen_text_digest, landing_page_text_digest, offer, observed_benefits, observed_conditions, account_ref, product_anchor_ref, landing_page_ref, scheduled_at, validated_at, receipt_digest]
      properties:
        schema_version: {const: contentcloud.douyin-commerce-validation/1.0}
        content_profile_id: {const: douyin-commerce-video}
        project_id: {type: string, minLength: 1}
        audience_strategy_approved_snapshot_id: {type: string, minLength: 1}
        audience_strategy_version_id: {type: string, minLength: 1}
        offer_approved_snapshot_id: {type: string, minLength: 1}
        offer_snapshot_id: {type: string, minLength: 1}
        content_approved_snapshot_id: {type: string, minLength: 1}
        content_item_id: {type: string, minLength: 1}
        content_item_digest: {$ref: '#/components/schemas/SHA256Digest'}
        storyboard_approved_snapshot_id: {type: string, minLength: 1}
        storyboard_package_id: {type: string, minLength: 1}
        storyboard_locked_digest: {$ref: '#/components/schemas/SHA256Digest'}
        rendered_creative_artifact_id: {type: string, minLength: 1}
        rendered_creative_digest: {$ref: '#/components/schemas/SHA256Digest'}
        voiceover_text_digest: {$ref: '#/components/schemas/SHA256Digest'}
        on_screen_text_digest: {$ref: '#/components/schemas/SHA256Digest'}
        landing_page_text_digest: {$ref: '#/components/schemas/SHA256Digest'}
        offer: {$ref: '#/components/schemas/DouyinCommerceOfferFacts'}
        observed_benefits: {type: array, uniqueItems: true, items: {type: string, minLength: 1}}
        observed_conditions: {type: array, uniqueItems: true, items: {type: string, minLength: 1}}
        account_ref: {type: string, minLength: 1}
        product_anchor_ref: {type: string, minLength: 1}
        landing_page_ref: {type: string, minLength: 1}
        scheduled_at: {type: string, format: date-time}
        validated_at: {type: string, format: date-time}
        receipt_digest: {$ref: '#/components/schemas/SHA256Digest'}
    DouyinCommercePublicationRefs:
      type: object
      additionalProperties: false
      required: [audience_strategy_approved_snapshot_id, audience_strategy_version_id, offer_approved_snapshot_id, offer_snapshot_id, content_approved_snapshot_id, content_item_id, storyboard_approved_snapshot_id, storyboard_package_id, rendered_creative_artifact_id, rendered_creative_digest, validation_receipt_digest, account_ref, product_anchor_ref, landing_page_ref, validation_receipt]
      properties:
        audience_strategy_approved_snapshot_id: {type: string, minLength: 1}
        audience_strategy_version_id: {type: string, minLength: 1}
        offer_approved_snapshot_id: {type: string, minLength: 1}
        offer_snapshot_id: {type: string, minLength: 1}
        content_approved_snapshot_id: {type: string, minLength: 1}
        content_item_id: {type: string, minLength: 1}
        storyboard_approved_snapshot_id: {type: string, minLength: 1}
        storyboard_package_id: {type: string, minLength: 1}
        rendered_creative_artifact_id: {type: string, minLength: 1}
        rendered_creative_digest: {$ref: '#/components/schemas/SHA256Digest'}
        validation_receipt_digest: {$ref: '#/components/schemas/SHA256Digest'}
        account_ref: {type: string, minLength: 1}
        product_anchor_ref: {type: string, minLength: 1}
        landing_page_ref: {type: string, minLength: 1}
        validation_receipt: {$ref: '#/components/schemas/DouyinCommerceValidationReceipt'}
    RecordManualChannelReceiptInput:
      type: object
      additionalProperties: false
      required: [state, external_id]
      properties:
        state: {type: string, enum: [submitted, published, failed, unknown, withdrawn]}
        external_id: {type: string}
        external_url: {type: string, format: uri}
        published_at: {type: string, format: date-time}
        error_code: {type: string}
        safe_summary: {type: object, additionalProperties: true}
...
    ReconcileChannelPublicationsInput:
      type: object
      additionalProperties: false
      properties:
        limit: {type: integer, minimum: 0, maximum: 1000, default: 100}
...
    SHA256Digest:
      type: string
      pattern: '^sha256:[a-f0-9]{64}$'
    SuccessEnvelope:
      type: object
      required: [ok, command, request_id, data, meta]
      properties:
        ok: {const: true}
        command: {type: string}
        request_id: {type: string}
        data: {}
        meta: {type: object}
    ErrorEnvelope:
      type: object
      required: [ok, command, request_id, error]
      properties:
        ok: {const: false}
        command: {type: string}
        request_id: {type: string}
        error:
          type: object
          required: [type, subtype, code, message, retryable]
          properties:
            type: {type: string}
            subtype: {type: string}
            code: {type: string}
            message: {type: string}
            retryable: {type: boolean}
            hint: {type: string}
            details: {}
  responses:
    Success:
      description: Stable success envelope
      content:
        application/json: {schema: {$ref: '#/components/schemas/SuccessEnvelope'}}
    Error:
      description: Stable structured error envelope
      content:
        application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}